+++ CHANGELOG.md
@@ -39,6 +39,11 @@
 ### 1.0.6
 - Change image loading placeholder
 
+### 1.2.1
+- [cinemagraph] Add cinemagraph
+- [fullpage-slides] Add fullpage-slides
+- Fix eslint imports resloving
+
 ### 1.0.5
 - Bug fix. Show tags on Category listing page.